requirements.txt
tqdm>=4.62.3
matplotlib>=3.5.3
numpy>=1.22.2
torch>=1.12
torchvision>=0.13
PyYAML>=6.0
